readme: add Using section
I believe this is what we intended, and it is what I want.

It could lead to duplicating low-level API in high-level API at first,
but once we start seeing DisplayID in the wild, the high-level functions
need to start looking into both EDID and DisplayID APIs to deliver the
desired information.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>